refactor : PR 내용 반영했습니다
