+++ src/api/mentor/server/getMentoringNewCount.ts
@@ -1,10 +1,6 @@
 import serverFetch from "@/utils/serverFetchUtil";
 
-// import { MentoringNewCountResponse } from "../type/response";
-
-export interface MentoringNewCountResponse {
-  count: number;
-}
+import { MentoringNewCountResponse } from "../type/response";
 
 const getMentoringNewCount = () => {
   return serverFetch<MentoringNewCountResponse>("/mentorings/check", {
