+++ src/utils/isServerStateLogin.ts
@@ -5,7 +5,7 @@ import { isTokenExpired } from "./jwtUtils";
 const isServerStateLogin = (): boolean => {
   const cookieStore = cookies();
   const refreshToken = cookieStore.get("refreshToken")?.value;
-  const isLogin = refreshToken ? isTokenExpired(refreshToken) : false;
+  const isLogin = !!(refreshToken && !isTokenExpired(refreshToken));
   return isLogin;
 };
 export default isServerStateLogin;
